renamed index.js to server.js, wrote integration tests to server.js
