+++ server/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "run-server": "cd ../client && npm run build && cd ../server && node index.js"
+    "launch-frontend": "cd ../client && npm run build && cd ../server && node index.js",
+    "test": "jest --detectOpenHandles"
   },
   "author": "",
   "license": "ISC",
@@ -13,6 +14,8 @@
     "@hapi/inert": "^6.0.3"
   },
   "devDependencies": {
-    "nodemon": "^2.0.12"
+    "jest": "^27.1.0",
+    "nodemon": "^2.0.12",
+    "supertest": "^6.1.6"
   }
 }
